+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-06 18:14:25
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-08-06 20:34:05
+// Commit: style(core): adjust design tokens, typography, and contrast in core
 
 class coreComponentState {
-  final String commitRef = '3411';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-08-06T18:14:25+05:00');
+  final String commitRef = '3424';
+  final String lastMessage = 'style(core): adjust design tokens, typography, and contrast in core';
+  final DateTime timestamp = DateTime.parse('2026-08-06T20:34:05+05:00');
 }